+++ docs/en/changes/changes.md
@@ -113,6 +113,7 @@
 * Update Grafana dashboards for OAP observability.
 * BanyanDB: fix query `getInstance` by instance ID.
 * Support the go agent(0.7.0 release) bundled pprof profiling feature. 
+* Service and TCPService source support analyze TLS mode.
 
 
 #### UI